A few bug fixes before I go to bed. V0.35

+++ main.py
@@ -14,7 +14,7 @@ from wx.lib.mixins.listctrl import ListCtrlAutoWidthMixin
 import wx.lib.mixins.listctrl as listmix
 from wx import wizard as wiz
 
-progVer = 0.3
+progVer = 0.35
 logfile = os.getcwd() + '/iwlist.log'
 
 euid = os.geteuid()
@@ -116,6 +116,7 @@ class WiFiz(wx.Frame):
 				                      "")
 		if self.UID.ShowModal() == wx.ID_OK:
 			self.UIDValue = self.UID.GetValue()
+		os.system("ifconfig " + self.UIDValue + " up")
 		self.OnScan(self)
 		
 	def OnConnect(self, e):
@@ -131,29 +132,38 @@ class WiFiz(wx.Frame):
 		workDir = "/etc/network.d/"
 		filename = str("wifiz-" + nameofProfile).strip()
 		filename = filename.strip()
-		f = open(workDir + filename, "w")
-		f.write("Description='A profile made by Wifiz for " + str(nameofProfile).strip() + "'\n")
-		f.close()
-		f = open(workDir + filename, 'a')
-		f.write("Interface=" + str(self.UIDValue).strip() + "\n")
-		f.write("Connection=wireless\n")
-		f.write("Security=" + typeofSecurity + "\n")
-		f.write("ESSID='" + str(nameofProfile).strip() + "'\n")
-		if str(typeofSecurity) != "None":  
-			passw = wx.TextEntryDialog(self, "What is the password?",
-				                              "Password",
-				                              "")
-			if passw.ShowModal() == wx.ID_OK:
-				thepass = passw.GetValue()	
+		if os.path.isfile(workDir + filename):
+			os.system("ifconfig " + self.UIDValue + " down")
+			os.system("netctl start wifiz-" + str(nameofProfile).strip())
+			self.OnScan(self)
+			wx.MessageBox("You are now connected to " + str(nameofProfile).strip() + ".", "Connected.")
+		else:
+			f = open(workDir + filename, "w")
+			f.write("Description='A profile made by Wifiz for " + str(nameofProfile).strip() + "'\n")
+			f.close()
+			f = open(workDir + filename, 'a')
+			f.write("Interface=" + str(self.UIDValue).strip() + "\n")
+			f.write("Connection=wireless\n")
+			f.write("Security=" + typeofSecurity + "\n")
+			f.write("ESSID='" + str(nameofProfile).strip() + "'\n")
+			if str(typeofSecurity).strip() != "none":  
+				passw = wx.TextEntryDialog(self, "What is the password?",
+					                              "Password",
+					                              "")
+				if passw.ShowModal() == wx.ID_OK:
+					thepass = passw.GetValue()	
+				else:
+					f.close()
+				f.write(r'Key=\"' + thepass + "\n")
 			else:
-				f.close()
-				break
-		f.write(r'Key=\"' + thepass + "\n")
-		f.write("IP=dhcp\n")
-		f.close()
-		os.system("netctl enable wifiz-" + str(nameofProfile).strip())
-		os.system("netctl start wifiz-" + str(nameofProfile).strip())
-		
+				f.write(r'Key=None\n')
+			f.write("IP=dhcp\n")
+			f.close()
+			os.system("ifconfig " + self.UIDValue + " down")
+			os.system("netctl enable wifiz-" + str(nameofProfile).strip())
+			os.system("netctl start wifiz-" + str(nameofProfile).strip())
+			wx.MessageBox("You are now connected to " + str(nameofProfile).strip() + ".", "Connected.")
+			self.OnScan(self)
 		
 	def getSelectedIndices( self, state =  wx.LIST_STATE_SELECTED):
 		indices = []
@@ -180,16 +190,14 @@ class WiFiz(wx.Frame):
 		self.APList.PopupMenu(self.PopupMenu, pos)
 
 	def OnDConnect(self, e):
-		print self.getSelectedIndices()
 		index = str(self.getSelectedIndices()).strip('[]')
 		index = int(index)
-		print index
 		item = self.APList.GetItem(index, 0)
 		nameofProfile = item.GetText()
-		print nameofProfile
-		wx.MessageBox(nameofProfile, nameofProfile)
-		os.system("sudo netctl stop " + nameofProfile)
-		
+		os.system("netctl stop wifiz-" + nameofProfile)
+		os.system("ifconfig " + self.UIDValue + " up")
+		self.OnScan(self)
+		wx.MessageBox("You are now disconnected from " + nameofProfile + ".", "Disconnected.")
 
 	def OnNew(self, e):
 		newProf = NewProfile(parent=None)       
@@ -221,7 +229,6 @@ class WiFiz(wx.Frame):
 				begin = line.replace(" ", "")
 				mid = begin.replace("ESSID:", "")
 				final = mid.replace('"', "")
-				print final
 				self.APList.SetStringItem(self.index, 0, final)
 				line = open(ilogfile).readline()
 				
@@ -240,7 +247,7 @@ class WiFiz(wx.Frame):
 				self.APList.SetStringItem(self.index, 1, s3)
 			if "Encryption" in line:
 				if "WPA2" in line:
-					encrypt = "WPA2"
+					encrypt = "WPA"
 				elif "off" in line:
 					encrypt = "None"
 				else:
